continued integration of collision engine
made collision engine control motion of the collisionobjects we give it
made rigidbody transform store a pointer to the mesh transform
made BoundingBox store a pointer to the mesh transform too.
Any transformations of a rigidbody will reflect as a transformation of the mesh, and any transformation of the mesh will reflect as a transformation of the bounding bod of the mesh.